Trash/ShamielaBookConverter/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shamela-books-converter"
version = "2.0.0"
description = "محول كتب الشاملة من Access إلى MySQL"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "GitHub Copilot AI Assistant"}
]
maintainers = [
    {name = "GitHub Copilot AI Assistant"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Natural Language :: Arabic",
]
keywords = ["shamela", "books", "converter", "access", "mysql", "database", "arabic"]
dependencies = [
    "pyodbc>=4.0.35",
    "mysql-connector-python>=8.0.33",
    "pyinstaller>=5.13.0",
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = ""
"Bug Reports" = ""
"Source" = ""
"Documentation" = ""

[project.scripts]
shamela-converter = "main:main"

[tool.setuptools]
packages = ["src"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.ico", "*.png", "*.json"]
"resources" = ["*.*"]